Correctly detect lack of scans
With 55feaf4cf7c714bddfb802bff3bc125118337c93, I changed the logic
to no longer always fetch and pass in scan data to the breach alert
email template, to avoid heavy database load for a query that in
most cases isn't necessary. With that change, I also changed the
check for whether a scan had been done to not look at the absence
of scan data, but to look at the absence of the data summary -- but
didn't pass that in *only* when no scan had occurred. Thus, it would
conclude that a scan had occurred in cases where it hadn't.

To fix this, this change ensures that we look at whether we have a
OneRep scan ID, which should only be set if a scan has actually
occurred.